+++ 3D課題/提出用/後期課題：3D/3DProgram/Src/Engine/Graphics/Drawer2D.h
@@ -11,8 +11,8 @@ class LineDesc;
 
 struct Texture{
 	LPDIRECT3DTEXTURE9 TexutreData;	//!< テクスチャデータ
-	int Width;						//!< 横幅
-	int Height;						//!< 縦幅
+	float Width;						//!< 横幅
+	float Height;						//!< 縦幅
 };
 
 struct CustomVertex {
